requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fastapi==0.89.1
idna==3.4
oauthlib==3.2.2
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
starlette==0.22.0
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
